+++ src/Firebase.php
@@ -4,6 +4,7 @@ namespace Kreait;
 
 use Kreait\Firebase\Auth;
 use Kreait\Firebase\Database;
+use Kreait\Firebase\Storage;
 
 class Firebase
 {
@@ -17,29 +18,30 @@ class Firebase
      */
     private $auth;
 
-    public function __construct(Database $database, Auth $auth)
+    /**
+     * @var Storage
+     */
+    private $storage;
+
+    public function __construct(Database $database, Auth $auth, Storage $storage)
     {
         $this->database = $database;
         $this->auth = $auth;
+        $this->storage = $storage;
     }
 
-    /**
-     * Returns an instance of the realtime database.
-     *
-     * @return Database
-     */
     public function getDatabase(): Database
     {
         return $this->database;
     }
 
-    /**
-     * Returns an Auth instance.
-     *
-     * @return Auth
-     */
     public function getAuth(): Auth
     {
         return $this->auth;
+    }
+
+    public function getStorage(): Storage
+    {
+        return $this->storage;
     }
 }